pyproject.toml
[project]
name = "haposa"
version = "0.1.0"
description = ""
authors = [
    {name = "tioluwa",email = "tioluwa.dev@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.2,<6.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
